updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -33,8 +33,10 @@ with app.app_context():
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user = db.session.query(User).filter_by(email='dbrow206@uncc.edu').one()
-    return render_template('index.html', user=a_user)
+
+    if session.get('user'):
+        return render_template("index.html", user=session['user'])
+    return render_template("index.html")
 
 
 @app.route('/notes')
@@ -50,58 +52,68 @@ def get_notes():
 
 @app.route('/notes/<note_id>')
 def get_note(note_id):
-    a_user = db.session.query(User).filter_by(email='dbrow206@uncc.edu').one()
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
+    if session.get('user'):
+        my_note = db.session.query(Note).filter_by(id=note_id, user_id=session['user_id']).one()
 
-    return render_template('note.html', note=my_note, user=a_user)
+        return render_template('note.html', note=my_note, user=session['user'])
+    else:
+        return redirect(url_for('login'))
 
 
 @app.route('/notes/new', methods=["GET", "POST"])
 def new_note():
-    if request.method == "POST":
-        title = request.form["title"]
-        text = request.form["noteText"]
-        from datetime import date
-        today = date.today()
-        today = today.strftime("%m-%d-%Y")
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
-        return redirect(url_for("get_notes"))
+    if session.get('user'):
 
+        if request.method == "POST":
+            title = request.form["title"]
+            text = request.form["noteText"]
+            from datetime import date
+            today = date.today()
+            today = today.strftime("%m-%d-%Y")
+            new_record = Note(title, text, today)
+            db.session.add(new_record)
+            db.session.commit()
+
+            return redirect(url_for("get_notes"))
+        else:
+            a_user = db.session.query(User).filter_by(id=session['user_id']).one()
+            return render_template('new.html', user=a_user)
     else:
-        a_user = db.session.query(User).filter_by(email='dbrow206@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+        return redirect(url_for('login'))
 
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
 
-        note.title = title
-        note.text = text
+            note.title = title
+            note.text = text
 
-        db.session.add(note)
-        db.session.commit()
-        return redirect(url_for('get_notes'))
+            db.session.add(note)
+            db.session.commit()
+            return redirect(url_for('get_notes'))
+        else:
+            a_user = db.session.query(User).filter_by(id=session['user_id']).one()
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+            return render_template('new.html', note=my_note, user=a_user)
 
     else:
-        a_user = db.session.query(User).filter_by(email='dbrow206@uncc.edu').one()
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note=my_note, user=a_user)
+        return redirect(url_for('login'))
 
 
 @app.route('/notes/delete/<note_id>', methods=['POST'])
 def delete_note(note_id):
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
-
-    return redirect((url_for('get_notes')))
+    if session.get('user'):
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
+        db.session.commit()
+        return redirect(url_for('get_notes'))
+    else:
+        return redirect((url_for('login')))
 
 
 @app.route('/register', methods=['POST', 'GET'])
